move test to integration dir
